Implemented the function which writes the results to an output file.
Note: When I compute the results, I need the original matrix T. Instead of
creating or reading it again, I decided to create a copy at the beginning,
which is much faster.

Additionally, at the moment I compute the matrix-vector product T*xi by
myself instead of calling a MKL function. This is due to the difference in
the packed format which is used in this BLAS2 routine (they expect one
array, while the LAPACK QR algorithm expects two).